Fixed maps not loading.
Turns out the ../../../wp-config.php include would not always work.
The fix was to use $_SERVER[DOCUMENT_ROOT] since wp_config.php is
always going to reside there anyway.   This was trigger because the
of a php file being used as a javascript include that sets vars later
needed by google maps based on the wp-config attributes.   What a mess.

Also, random code cleanup (formatting) while searching for the source
of the bug.  I know... bad way to do a commit. :/