Fixd map lookup feature.  Was not returning results.
Turns out the generate-xml.php includes wp-config.php, which then
loads up all the plugin files.   This is probably a bad way to do
this and incurs a ton of extra overhead.   We need to find a way
to get the DB connection info out of wp-config or at least disable
plugin calls.

This broke because SLPLUS_PLUGINDIR was not defined, breaking our
standard include/config.php.