Make map hide image & show map on location search.
You can now specify a starting static image which is hidden when
a search is performed.  The store-locator.js uses jqeury and some
Google Maps API trickery to show the map and then resize the map
to ensure proper rendering (resize is necessary else the GM gets
confused as to its real size).

Redmine-Ticket: 4053