Add the "SLP Editor" role for users = editor with manage SLP locations.
Also added the manage_slp capability to the admin user.

For SLP Editor role to work the WPCSL-generic system needs to be updated to not just check for current_user_can = administrator.  We should be checking that he user can manage plugins v. just be a system admin.

As such SLP Editor role is coming out on the very next commit until that is fixed.