Update config file to manage plus packs added on to functional base product.
This is the first step in getting the Plus Edition to replace the LE version on WordPress.  We are moving the installation of a complete plus product into a paid add-on-features package called the "Plus Pack".   This simply sets up the classes and payment identifiers.  Next we need to wire in the license lookup and only enable certain features when a plus pack license is turned on (currently all the features are enabled all the time, like reporting).