Create an icon to show which locations have a page.
Check if a page already exists, if so update it v. create new one.

Check if a page exists at all, they can be perm deleted via another channel, if so don't barf, instead just recreate a new page when the update routine is called.

Change the message text to tell people things are updating or creating.

Ticket: #7881
Ticket: #7882
Ticket: #7884