Attempt to fix interim JS problem.
Google main.js is barfing on  geocoder.geocode() with an "a is null" error.

I think it has something to do with re-declaring this to _this and passing it to a separate function. The overhead is slow on some sites and causes the object stack to not be fully intact when subsequent processing occurs.