Prevent duplicate form entries.
For some reason various themes (or another plugin?) is calling the div filters TWICE.
I confirmed that the fitler is only being added once and is only in the filter stack once.
Calling remove filter in the processing loop does not help.
It appears that if the filter changes the content of the incoming string the filter is being re-applied.  No idea what is triggering that behavior but it has been reported at least.... umm... twice.  :/