Wrap all error_log debugging in a test for Debug My Plugin being active.
Need to use error_log because some of the WP pages for this stuff happen outside the normal UI interface and thus Debug Bar/Debug My Plugin do not work as you'd expect.

However, we want to use the Debug My Plugin hook to control all debugging on/off control, so we test for that.