Add options property to SLPlus.
This is a single serialized option that is read from the DB.  Cuts down on data I/O, should be used for ALL new core plugin options.

This also gets passed as a single localized var to the csl.js script, making all main plugin options readily available to our script for less script localization hacking.

Too bad this was not designed this way originally.